안쓰는 파일 정리 및 서버 연결을 위한 네트워크 로직 삽입

+++ GameCoding/GameCoding/GameMonster.cpp
@@ -138,7 +138,7 @@ void GameMonster::TickSkill()
 	if (_waitSeconds > 0)
 	{
 		float deltaTime = GET_SINGLE(TimeManager)->GetDeltaTime();
-		_waitSeconds = max(0, _waitSeconds - deltaTime);
+		_waitSeconds = max(0, static_cast<int32>(_waitSeconds - deltaTime));
 		return;
 	}
